+++ agrest-openapi/src/test/java/io/agrest/openapi/ResponseTest.java
@@ -2,17 +2,19 @@ package io.agrest.openapi;
 
 import io.agrest.DataResponse;
 import io.agrest.entity.AgP1;
+import io.agrest.openapi.unit.OpenAPIBuilder;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
-import org.example.entity.P1;
-import io.agrest.openapi.unit.OpenAPIBuilder;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.PathItem;
 import io.swagger.v3.oas.models.media.MediaType;
 import io.swagger.v3.oas.models.media.Schema;
 import io.swagger.v3.oas.models.responses.ApiResponse;
 import io.swagger.v3.oas.models.responses.ApiResponses;
+import org.example.entity.P1;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
@@ -70,9 +72,10 @@ public class ResponseTest {
         assertEquals("#/components/schemas/DataResponse(AgP1)", schema.get$ref());
     }
 
-    @Test
-    public void testDataResponseWithNoType() {
-        PathItem pi = oapi.getPaths().get("/r/data-response-bare");
+    @ParameterizedTest
+    @ValueSource(strings = {"/r/data-response-bare", "/r/data-response-wildcard"})
+    public void testDataResponseWithNoType(String uri) {
+        PathItem pi = oapi.getPaths().get(uri);
         ApiResponses responses = pi.getGet().getResponses();
 
         assertEquals(1, responses.size());
@@ -131,6 +134,12 @@ public class ResponseTest {
         @GET
         @Path("data-response-bare")
         public DataResponse dataResponseBare() {
+            throw new UnsupportedOperationException("endpoint logic is irrelevant for the test");
+        }
+
+        @GET
+        @Path("data-response-wildcard")
+        public DataResponse<?> dataResponseWildcard() {
             throw new UnsupportedOperationException("endpoint logic is irrelevant for the test");
         }