+++ agrest-openapi/src/test/java/org/example/entity/P8.java
@@ -1,0 +1,16 @@
+package org.example.entity;
+
+import io.agrest.annotation.AgAttribute;
+
+public class P8 {
+
+    @AgAttribute
+    public String getOtherP8() {
+        return null;
+    }
+
+    @AgAttribute
+    public String getDuplicate() {
+        return null;
+    }
+}