+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## Release 4.1
 
+* #459 "code-first" OpenAPI support and Swagger integration
 * #460 Support for the new Cayenne "Json" type
 * #465 AgIdPart metadata object